+++ src/components/Dialogs/PreferenceDialog/PreferenceDialogComponent.tsx
@@ -12,23 +12,7 @@ import tinycolor from "tinycolor2";
 
 import {DraggableDialogComponent} from "components/Dialogs";
 import {AppToaster, AutoColorPickerComponent, ColormapComponent, ColorPickerComponent, PointShapeSelectComponent, SafeNumericInput, ScalingSelectComponent, SuccessToast} from "components/Shared";
-import {
-    CompressionQuality,
-    CursorInfoVisibility,
-    CursorPosition,
-    Event,
-    FileFilterMode,
-    RegionCreationMode,
-    SPECTRAL_MATCHING_TYPES,
-    SPECTRAL_TYPE_STRING,
-    SpectralType,
-    Theme,
-    TileCache,
-    WCSMatchingType,
-    WCSType,
-    Zoom,
-    ZoomPoint
-} from "models";
+import {CompressionQuality, CursorInfoVisibility, CursorPosition, Event, FileFilterMode, RegionCreationMode, SPECTRAL_MATCHING_TYPES, SPECTRAL_TYPE_STRING, Theme, TileCache, WCSMatchingType, WCSType, Zoom, ZoomPoint} from "models";
 import {TelemetryMode} from "services";
 import {AppStore, BeamType, HelpType, PreferenceKeys, PreferenceStore} from "stores";
 import {ContourGeneratorType, FrameScaling, RegionStore, RenderConfigStore} from "stores/Frame";
@@ -47,7 +31,8 @@ enum PreferenceDialogTabs {
     PERFORMANCE,
     LOG_EVENT,
     CATALOG,
-    TELEMETRY
+    TELEMETRY,
+    COMPATIBILITY
 }
 
 export enum MemoryUnit {
@@ -145,6 +130,9 @@ export class PreferenceDialogComponent extends React.Component {
             case PreferenceDialogTabs.TELEMETRY:
                 preference.resetTelemetrySettings();
                 break;
+            case PreferenceDialogTabs.COMPATIBILITY:
+                preference.resetCompatibilitySettings();
+                break;
             case PreferenceDialogTabs.GLOBAL:
             default:
                 preference.resetGlobalSettings();
@@ -228,7 +216,7 @@ export class PreferenceDialogComponent extends React.Component {
                     </HTMLSelect>
                 </FormGroup>
                 <FormGroup inline={true} label="Spectral matching">
-                    <HTMLSelect value={preference.spectralMatchingType} onChange={ev => appStore.setSpectralMatchingType(ev.currentTarget.value as SpectralType)}>
+                    <HTMLSelect value={preference.spectralMatchingType} onChange={ev => preference.setPreference(PreferenceKeys.GLOBAL_SPECTRAL_MATCHING_TYPE, ev.currentTarget.value)}>
                         {SPECTRAL_MATCHING_TYPES.map(type => (
                             <option key={type} value={type}>
                                 {SPECTRAL_TYPE_STRING.get(type)}
@@ -874,6 +862,14 @@ export class PreferenceDialogComponent extends React.Component {
             </div>
         );
 
+        const compatibilityPanel = (
+            <div className="panel-container">
+                <FormGroup inline={true} label="AIPS cube beam support">
+                    <Switch checked={preference.aipsBeamSupport} onChange={ev => preference.setPreference(PreferenceKeys.COMPATIBILITY_AIPS_BEAM_SUPPORT, ev.currentTarget.checked)} />
+                </FormGroup>
+            </div>
+        );
+
         const className = classNames("preference-dialog", {"bp3-dark": appStore.darkTheme});
 
         const dialogProps: IDialogProps = {
@@ -901,6 +897,7 @@ export class PreferenceDialogComponent extends React.Component {
                         <Tab id={PreferenceDialogTabs.ANNOTATION} title="Annotation" panel={annotationSettingsPanel} />
                         <Tab id={PreferenceDialogTabs.PERFORMANCE} title="Performance" panel={performancePanel} />
                         {process.env.REACT_APP_SKIP_TELEMETRY !== "true" && <Tab id={PreferenceDialogTabs.TELEMETRY} title="Telemetry" panel={telemetryPanel} />}
+                        <Tab id={PreferenceDialogTabs.COMPATIBILITY} title="Compatibility" panel={compatibilityPanel} />
                         <Tab id={PreferenceDialogTabs.LOG_EVENT} title="Log Events" panel={logEventsPanel} />
                     </Tabs>
                 </div>