+++ src/models/ChannelInfo/ChannelInfo.ts
@@ -1,9 +1,7 @@
 export interface ChannelInfo {
     fromWCS: boolean;
-    delta: number;
     indexes: number[];
     values: number[];
-    rawValues: number[];
     getChannelIndexWCS: (x: number) => number;
     getChannelIndexSimple: (x: number) => number;
 }