+++ src/stores/AppStore/AppStore.ts
@@ -71,6 +71,7 @@ interface ViewUpdate {
     channel: number;
     stokes: number;
     focusPoint: Point2D;
+    headerUnit: string;
 }
 
 interface ChannelUpdate {
@@ -347,6 +348,12 @@ export class AppStore {
         }
     }
 
+    // Spectral matching type, initialized by global preferences, modified by the Image List Settings
+    @observable spectralMatchingType: SpectralType;
+
+    // Match generated moment image(s) to the spatial reference image
+    @observable momentToMatch: boolean;
+
     @computed get openFileDisabled(): boolean {
         return this.backendService?.connectionStatus !== ConnectionStatus.ACTIVE || this.fileLoading;
     }
@@ -444,22 +451,25 @@ export class AppStore {
 
     @computed get spatialAndSpectalMatchedFileIds(): number[] {
         let matchedIds = [];
-        if (this.spatialReference) {
-            matchedIds.push(this.spatialReference.frameInfo.fileId);
-        }
+        const spatialReferenceId = this.spatialReference?.frameInfo?.fileId;
+        const spectralReferenceId = this.spectralReference?.frameInfo?.fileId;
 
         const spatialMatchedFileIds = this.spatialReference?.spatialSiblings?.map(matchedFrame => {
             return matchedFrame.frameInfo.fileId;
         });
-        const spectralMatchedFileIds = this.spatialReference?.spectralSiblings?.map(matchedFrame => {
+        spatialMatchedFileIds?.unshift(spatialReferenceId);
+        const spectralMatchedFileIds = this.spectralReference?.spectralSiblings?.map(matchedFrame => {
             return matchedFrame.frameInfo.fileId;
         });
+        spectralMatchedFileIds?.unshift(spectralReferenceId);
+
         spatialMatchedFileIds?.forEach(spatialMatchedFileId => {
             if (spectralMatchedFileIds?.includes(spatialMatchedFileId)) {
                 matchedIds.push(spatialMatchedFileId);
             }
         });
-        return matchedIds;
+
+        return [...new Set(matchedIds)]; //Remove duplicate
     }
 
     // Calculates which frames have a contour visible as a function of each visible frame
@@ -482,7 +492,19 @@ export class AppStore {
         return frameMap;
     }
 
-    @action addFrame = (ack: CARTA.IOpenFileAck, directory: string, lelExpr: boolean, hdu: string, generated: boolean = false, setAsActive: boolean = true): boolean => {
+    /**
+     * Adds a frame to the frame array based on the provided parameters.
+     *
+     * @param ack - The ack message received after opening a file.
+     * @param directory - The path to the parent directory of the file.
+     * @param lelExpr - Whether the file is opened with an image arithmetic (CASA lattice expression) string.
+     * @param hdu - The Header Data Unit (HDU) identifier of the file.
+     * @param generated - Whether the frame is a generated in-memory image. Used for the telemetry message.
+     * @param setAsActive - Whether the frame should be set as the active frame.
+     * @param updateStartingDirectory - Whether to update the starting directory in the file browser. Required for carta-python.
+     * @returns Whether the frame was successfully added.
+     */
+    @action addFrame = (ack: CARTA.IOpenFileAck, directory: string, lelExpr: boolean, hdu: string, generated: boolean = false, setAsActive: boolean = true, updateStartingDirectory: boolean = true): boolean => {
         if (!ack) {
             return false;
         }
@@ -554,7 +576,10 @@ export class AppStore {
                 this.setSpectralMatchingEnabled(newFrame, true);
             }
         }
-        this.fileBrowserStore.saveStartingDirectory(newFrame.frameInfo.directory);
+
+        if (updateStartingDirectory) {
+            this.fileBrowserStore.saveStartingDirectory(newFrame.frameInfo.directory);
+        }
 
         return true;
     };
@@ -590,8 +615,21 @@ export class AppStore {
         return newFrame;
     };
 
+    /**
+     * Loads a file at the given path and adds it as a frame to the application.
+     *
+     * @param path - The path to the parent directory of the file to load, or of the file itself.
+     * @param filename - The filename of the file to load.
+     * @param hdu - The Header Data Unit (HDU) to load. If left blank, the first image HDU will be loaded.
+     * @param imageArithmetic - Whether to treat the file as an image arithmetic (CASA lattice expression) string.
+     * @param setAsActive - Whether the loaded frame should be set as the active frame.
+     * @param updateStartingDirectory - Whether to update the starting directory in the file browser. Required for carta-python.
+     * @returns The added frame.
+     *
+     * @throws If there is an error loading the file.
+     */
     @flow.bound
-    *loadFile(path: string, filename: string, hdu: string, imageArithmetic: boolean, setAsActive: boolean = true) {
+    *loadFile(path: string, filename: string, hdu: string, imageArithmetic: boolean, setAsActive: boolean = true, updateStartingDirectory: boolean = true) {
         this.startFileLoading();
 
         if (imageArithmetic) {
@@ -611,7 +649,7 @@ export class AppStore {
         }
 
         // Separate HDU and filename if no HDU is specified
-        if (!hdu?.length) {
+        if (!hdu?.length && !imageArithmetic) {
             const hduRegex = /^(.*)\[(\S+)]$/;
             const matches = hduRegex.exec(filename);
             // Three matching groups. Second is filename, third is HDU
@@ -624,7 +662,7 @@ export class AppStore {
         try {
             const ack = yield this.backendService.loadFile(path, filename, hdu, this.fileCounter, imageArithmetic);
             this.fileCounter++;
-            if (!this.addFrame(ack, path, imageArithmetic, hdu, setAsActive)) {
+            if (!this.addFrame(ack, path, imageArithmetic, hdu, false, setAsActive, updateStartingDirectory)) {
                 AppToaster.show({icon: "warning-sign", message: "Load file failed.", intent: "danger", timeout: 3000});
             }
             this.endFileLoading();
@@ -677,34 +715,45 @@ export class AppStore {
     };
 
     /**
-     * Appends a file at the given path to the list of existing open files
-     * @access public
-     * @async
-     * @param path - path to the parent directory of the file to open, or of the file itself
-     * @param {string=} filename - filename of the file to open
-     * @param {string=} hdu - HDU to open. If left blank, the first image HDU will be opened
-     * @param {boolean=} imageArithmetic - Whether to treat the filename as an image arithmetic (CASA lattice expression) string
-     * @return {Promise<FrameStore>} [async] the FrameStore the opened file
+     * Appends a file at the given path to the list of existing open files.
+     *
+     * @param path - The path to the parent directory of the file to open, or of the file itself.
+     * @param filename - The filename of the file to open.
+     * @param hdu - The Header Data Unit (HDU) to open. If left blank, the first image HDU will be opened.
+     * @param imageArithmetic - Whether to treat the filename as an image arithmetic (CASA lattice expression) string.
+     * @param setAsActive - Whether to set the appended file as the active frame.
+     * @param updateStartingDirectory - Whether to update the starting directory in the file browser. Required for carta-python.
+     * @returns A promise that resolves to the FrameStore of the opened file.
+     *
+     * @example
+     * // Append a file with the given path and filename
+     * const file = await appendFile("/path/to/directory", "example.fits");
      */
     @flow.bound
-    *appendFile(path: string, filename?: string, hdu?: string, imageArithmetic: boolean = false, setAsActive: boolean = true) {
+    *appendFile(path: string, filename?: string, hdu?: string, imageArithmetic: boolean = false, setAsActive: boolean = true, updateStartingDirectory: boolean = true) {
         // Stop animations playing before loading a new frame
         this.animatorStore.stopAnimation();
-        return yield this.loadFile(path, filename, hdu, imageArithmetic, setAsActive);
+        return yield this.loadFile(path, filename, hdu, imageArithmetic, setAsActive, updateStartingDirectory);
     }
 
     /**
-     * Closes all existing files and opens a file at the given path
-     * @param path - path to the parent directory of the file to open, or of the file itself
-     * @param {string=} filename - filename of the file to open
-     * @param {string=} hdu - HDU to open. If left blank, the first image HDU will be opened
-     * @param {boolean=} imageArithmetic - Whether to treat the filename as an image arithmetic (CASA lattice expression) string
-     * @return {Promise<FrameStore>} [async] the FrameStore of the opened file
+     * Closes all existing files and opens a file at the given path.
+     *
+     * @param path - The path to the parent directory of the file to open, or of the file itself.
+     * @param filename - The filename of the file to open.
+     * @param hdu - The Header Data Unit (HDU) to open. If left blank, the first image HDU will be opened.
+     * @param imageArithmetic - Whether to treat the filename as an image arithmetic (CASA lattice expression) string.
+     * @param updateStartingDirectory - Whether to update the starting directory in the file browser. Required for carta-python.
+     * @returns A promise that resolves to the FrameStore of the opened file.
+     *
+     * @example
+     * // Open a file with the given path and filename
+     * const openedFile = await openFile("/path/to/directory", "example.fits");
      */
     @flow.bound
-    *openFile(path: string, filename?: string, hdu?: string, imageArithmetic?: boolean) {
+    *openFile(path: string, filename?: string, hdu?: string, imageArithmetic?: boolean, updateStartingDirectory: boolean = true) {
         this.removeAllFrames();
-        return yield this.loadFile(path, filename, hdu, imageArithmetic);
+        return yield this.loadFile(path, filename, hdu, imageArithmetic, true, updateStartingDirectory);
     }
 
     @flow.bound
@@ -750,8 +799,9 @@ export class AppStore {
     }
 
     /**
-     * Closes the currently active image
-     * @param confirmClose [boolean=] - Flag indicating whether to display a confirmation dialog before closing
+     * Closes the currently active image.
+     *
+     * @param confirmClose - Flag indicating whether to display a confirmation dialog before closing.
      */
     @action closeCurrentFile = (confirmClose: boolean = false) => {
         if (!this.appendFileDisabled) {
@@ -1186,7 +1236,11 @@ export class AppStore {
                 for (const openFileAck of ack.openFileAcks) {
                     if (this.addFrame(CARTA.OpenFileAck.create(openFileAck), this.fileBrowserStore.startingDirectory, false, "", true)) {
                         this.fileCounter++;
-                        frame.addMomentImage(this.frames.find(f => f.frameInfo.fileId === openFileAck.fileId));
+                        const newMomentImage = this.frames.find(f => f.frameInfo.fileId === openFileAck.fileId);
+                        frame.addMomentImage(newMomentImage);
+                        if (frame === this.spatialReference && this.momentToMatch) {
+                            newMomentImage.setSpatialReference(this.spatialReference);
+                        }
                     } else {
                         AppToaster.show({icon: "warning-sign", message: "Load file failed.", intent: "danger", timeout: 3000});
                     }
@@ -1485,7 +1539,11 @@ export class AppStore {
                 // TODO: dynamic tile size
                 const tileSizeFullRes = reqView.mip * 256;
                 const midPointTileCoords = {x: midPointImageCoords.x / tileSizeFullRes - 0.5, y: midPointImageCoords.y / tileSizeFullRes - 0.5};
-                this.tileService.requestTiles(tiles, frame.frameInfo.fileId, frame.channel, frame.stokes, midPointTileCoords, this.preferenceStore.imageCompressionQuality, true);
+
+                // If BUNIT = km/s, adopted compressionQuality is set to 32 regardless the preferences setup
+                const bunitVariant = ["km/s", "km s-1", "km s^-1", "km.s-1"];
+                const compressionQuality = bunitVariant.includes(frame.headerUnit) ? Math.max(this.preferenceStore.imageCompressionQuality, 32) : this.preferenceStore.imageCompressionQuality;
+                this.tileService.requestTiles(tiles, frame.frameInfo.fileId, frame.channel, frame.stokes, midPointTileCoords, compressionQuality, true);
             } else {
                 this.tileService.updateHiddenFileChannels(frame.frameInfo.fileId, frame.channel, frame.stokes);
             }
@@ -1520,11 +1578,11 @@ export class AppStore {
 
     private updateViews = (updates: ViewUpdate[]) => {
         for (const update of updates) {
-            this.updateView(update.tiles, update.fileId, update.channel, update.stokes, update.focusPoint);
+            this.updateView(update.tiles, update.fileId, update.channel, update.stokes, update.focusPoint, update.headerUnit);
         }
     };
 
-    private updateView = (tiles: TileCoordinate[], fileId: number, channel: number, stokes: number, focusPoint: Point2D) => {
+    private updateView = (tiles: TileCoordinate[], fileId: number, channel: number, stokes: number, focusPoint: Point2D, headerUnit: string) => {
         const isAnimating = this.animatorStore.serverAnimationActive;
         if (isAnimating) {
             this.backendService.addRequiredTiles(
@@ -1533,7 +1591,10 @@ export class AppStore {
                 this.preferenceStore.animationCompressionQuality
             );
         } else {
-            this.tileService.requestTiles(tiles, fileId, channel, stokes, focusPoint, this.preferenceStore.imageCompressionQuality);
+            // If BUNIT = km/s, adopted compressionQuality is set to 32 regardless the preferences setup
+            const bunitVariant = ["km/s", "km s-1", "km s^-1", "km.s-1"];
+            const compressionQuality = bunitVariant.includes(headerUnit) ? Math.max(this.preferenceStore.imageCompressionQuality, 32) : this.preferenceStore.imageCompressionQuality;
+            this.tileService.requestTiles(tiles, fileId, channel, stokes, focusPoint, compressionQuality);
         }
     };
 
@@ -1557,6 +1618,7 @@ export class AppStore {
                 await this.loadDefaultFiles();
                 this.setCursorFrozen(this.preferenceStore.isCursorFrozen);
                 this.updateASTColors();
+                this.setSpectralMatchingType(this.preferenceStore.spectralMatchingType);
                 if (this.preferenceStore.checkNewRelease) {
                     await this.checkNewRelease();
                 }
@@ -1628,6 +1690,7 @@ export class AppStore {
         this.toolbarExpanded = true;
         this.imageRatio = 1;
         this.isExportingImage = false;
+        this.momentToMatch = true;
 
         AST.onReady.then(
             action(() => {
@@ -1666,6 +1729,15 @@ export class AppStore {
             }
         }
         this.handleThemeChange(mediaQuery.matches);
+
+        // Add spellcheck="false" in input elements with no spellcheck attribute
+        document.addEventListener("input", (ev: Event) => {
+            const target = ev.target;
+            const isInput = target && target instanceof HTMLInputElement;
+            if (isInput && !target.getAttribute("spellcheck")) {
+                target.setAttribute("spellcheck", "false");
+            }
+        });
 
         // Display toasts when connection status changes
         autorun(() => {
@@ -1728,9 +1800,19 @@ export class AppStore {
                     const tileSizeFullRes = reqView.mip * 256;
                     const midPointTileCoords = {x: midPointImageCoords.x / tileSizeFullRes - 0.5, y: midPointImageCoords.y / tileSizeFullRes - 0.5};
                     if (tiles.length) {
-                        viewUpdates.push({tiles, fileId: frame.frameInfo.fileId, channel: frame.channel, stokes: frame.stokes, focusPoint: midPointTileCoords});
+                        viewUpdates.push({tiles, fileId: frame.frameInfo.fileId, channel: frame.channel, stokes: frame.stokes, focusPoint: midPointTileCoords, headerUnit: frame.headerUnit});
                     }
                 }
+
+                // Clear tiles of invisible matched images during animation
+                if (this.animatorStore?.serverAnimationActive) {
+                    for (const frame of this.activeFrame.spectralSiblings) {
+                        if (!this.visibleFrames.includes(frame)) {
+                            viewUpdates.push({tiles: [], fileId: frame.frameInfo.fileId, channel: frame.channel, stokes: frame.stokes, focusPoint: null, headerUnit: frame.headerUnit});
+                        }
+                    }
+                }
+
                 throttledSetViews(viewUpdates);
             }
         });
@@ -1738,26 +1820,30 @@ export class AppStore {
         // TODO: Move setChannels actions to AppStore and remove this autorun
         // Update channels when manually changed
         autorun(() => {
-            if (this.activeFrame) {
-                const updates: ChannelUpdate[] = [];
-                // Calculate if new data is required for the active channel
-                const updateRequiredChannels = this.activeFrame.requiredChannel !== this.activeFrame.channel || this.activeFrame.requiredStokes !== this.activeFrame.stokes;
-                // Don't auto-update when animation is playing
-                if (!this.animatorStore.animationActive && updateRequiredChannels) {
-                    updates.push({frame: this.activeFrame, channel: this.activeFrame.requiredChannel, stokes: this.activeFrame.requiredStokes});
-                }
+            const updates: ChannelUpdate[] = [];
 
-                // Update any sibling channels
-                this.activeFrame.spectralSiblings.forEach(frame => {
-                    const siblingUpdateRequired = frame.requiredChannel !== frame.channel || frame.requiredStokes !== frame.stokes;
-                    if (siblingUpdateRequired) {
-                        updates.push({frame, channel: frame.requiredChannel, stokes: frame.requiredStokes});
+            for (const visibleFrame of this.visibleFrames) {
+                if (visibleFrame) {
+                    // Calculate if new data is required for the active channel
+                    const updateRequiredChannels = visibleFrame.requiredChannel !== visibleFrame?.channel || visibleFrame.requiredStokes !== visibleFrame.stokes;
+                    // Don't auto-update when animation is playing
+                    if (!this.animatorStore.animationActive && updateRequiredChannels) {
+                        updates.push({frame: visibleFrame, channel: visibleFrame.requiredChannel, stokes: visibleFrame.requiredStokes});
                     }
-                });
 
-                if (updates.length) {
-                    throttledSetChannels(updates);
+                    // Update any sibling channels
+                    visibleFrame.spectralSiblings.forEach(frame => {
+                        const isVisible = this.visibleFrames.includes(frame);
+                        const siblingUpdateRequired = frame.requiredChannel !== frame.channel || frame.requiredStokes !== frame.stokes;
+                        if (!isVisible && siblingUpdateRequired) {
+                            updates.push({frame, channel: frame.requiredChannel, stokes: frame.requiredStokes});
+                        }
+                    });
                 }
+            }
+
+            if (updates.length) {
+                throttledSetChannels(updates);
             }
         });
 
@@ -1909,7 +1995,7 @@ export class AppStore {
     };
 
     @action handleTileStream = (tileStreamDetails: TileStreamDetails) => {
-        if (this.animatorStore.serverAnimationActive) {
+        if (this.animatorStore.serverAnimationActive && tileStreamDetails?.fileId === this.activeFrameFileId) {
             const frame = this.getFrame(tileStreamDetails.fileId);
             // Flow control
             const flowControlMessage: CARTA.IAnimationFlowControl = {
@@ -2156,7 +2242,8 @@ export class AppStore {
                 stokes: frame.requiredStokes,
                 regions: mapToObject(regions),
                 contourSettings,
-                stokesFiles: frame.stokesFiles
+                stokesFiles: frame.stokesFiles,
+                supportAipsBeam: AppStore.Instance.preferenceStore.aipsBeamSupport
             };
         });
 
@@ -2706,8 +2793,8 @@ export class AppStore {
         this.setSpectralMatchingEnabled(frame, !frame.spectralReference);
     };
 
-    setSpectralMatchingType = (spectralMatchingType: SpectralType) => {
-        this.preferenceStore.setPreference(PreferenceKeys.GLOBAL_SPECTRAL_MATCHING_TYPE, spectralMatchingType);
+    @action setSpectralMatchingType = (spectralMatchingType: SpectralType) => {
+        this.spectralMatchingType = spectralMatchingType;
         for (const f of this.frames) {
             if (f.spectralReference) {
                 this.setSpectralMatchingEnabled(f, true);
@@ -2768,6 +2855,10 @@ export class AppStore {
     @action setMatchingEnabled = (spatial: boolean, spectral: boolean) => {
         this.setSpatialMatchingEnabled(this.activeFrame, spatial);
         this.setSpectralMatchingEnabled(this.activeFrame, spectral);
+    };
+
+    @action toggleMomentToMatch = () => {
+        this.momentToMatch = !this.momentToMatch;
     };
 
     @computed get numImagePages() {